refactor: rename ui
