+++ src/ui/mod.rs
@@ -1,2 +1,2 @@
+pub mod main_ui;
 pub mod popups;
-pub mod ui;
